refactor(reconciliation): split balance assertion modules
Why:
- The balance assertion implementation had grown past the repo's 300-line refactor limit and mixed row models with issue-assembly logic in one module.

What:
- Split assertion formatting, models, and comparison logic into separate package modules under the existing reconciliation seam.
- Remove the oversized single-module implementation without changing the package-level API.

Checks:
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pytest -q --no-cov tests/unit/domain/reconciliation/test_assertions.py tests/unit/application/reconciliation/test_assert_balances.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run ruff check src/tallylot/domain/reconciliation tests/unit/domain/reconciliation/test_assertions.py
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run mypy src/tallylot/domain/reconciliation
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pyright src/tallylot/domain/reconciliation
- UV_PROJECT_ENVIRONMENT="/home/user/.venvs/tallylot-py312" uv run pylint src/tallylot/domain/reconciliation
