+++ src/tallylot/domain/reconciliation/__init__.py
@@ -1,9 +1,11 @@
 """Provider-neutral reconciliation seams."""
 
-from .assertions import (
+from .assertion_models import (
     BalanceAssertion,
     BalanceAssertionResult,
     BalanceAssertionStatus,
+)
+from .assertion_service import (
     assert_balance_snapshots,
 )
 from .evidence import BalanceEvidence
